@@ -17,7 +17,7 @@
     ]
   },
   "require": {
-    "php": ">=5.6",
+    "php": ">=8.1",
     "ext-mbstring": "*"
   }
 }
