@@ -1,13 +1,13 @@
 <?php
+declare(strict_types=1);
 
 use Manuylenko\Dumper\Dumper;
 
-
 if (! function_exists('dump')) {
     /**
-     * @return void
+     * ..
      */
-    function dump() {
+    function dump(): void {
         foreach (func_get_args() as $var) {
             (new Dumper())->dump($var);
         }
@@ -16,9 +16,9 @@ if (! function_exists('dump')) {
 
 if (! function_exists('dumpEx')) {
     /**
-     * @return void
+     * ..
      */
-    function dumpEx() {
+    function dumpEx(): never {
         call_user_func_array('dump', func_get_args());
         exit;
     }
