@@ -1,15 +1,14 @@
 <?php
+declare(strict_types=1);
 
 namespace Manuylenko\Dumper\Types;
 
 class BooleanType extends Type
 {
     /**
-     * @param bool $bool
-     *
-     * @return string
+     * ..
      */
-    public static function render($bool)
+    public static function render(bool $bool): string
     {
         return '<span class="md_boolean" title="Тип: bool">'.($bool ? 'true' : 'false').'</span>';
     }
